+++ src/main/kotlin/me/parker/nextstep/kotlinracingcar/controller/RacingGameController.kt
@@ -1,7 +1,7 @@
 package me.parker.nextstep.kotlinracingcar.controller
 
-import me.parker.nextstep.kotlinracingcar.RacingGame
-import me.parker.nextstep.kotlinracingcar.rule.RandomRacingGameMoveRule
+import me.parker.nextstep.kotlinracingcar.domain.RacingGame
+import me.parker.nextstep.kotlinracingcar.domain.rule.RandomRacingGameMoveRule
 import me.parker.nextstep.kotlinracingcar.view.ConsoleInput
 import me.parker.nextstep.kotlinracingcar.view.ConsoleResult
 
@@ -10,7 +10,8 @@ fun main(args: Array<String>) {
     val numOfAttempts: Int = ConsoleInput.inputNumOfAttempts()
 
     val racingGame = RacingGame(namesOfRacingCar.split(ConsoleInput.NAME_DELIMITER),
-        numOfAttempts, RandomRacingGameMoveRule())
+        numOfAttempts, RandomRacingGameMoveRule()
+    )
     val racingGameResult = racingGame.start()
 
     ConsoleResult.outputGameResult(racingGameResult)
