+++ src/main/kotlin/me/parker/nextstep/kotlinracingcar/domain/RacingCarGameRound.kt
@@ -1,4 +1,4 @@
-package me.parker.nextstep.kotlinracingcar
+package me.parker.nextstep.kotlinracingcar.domain
 
 data class RacingCarGameRound(val roundNumber: Int, val currentRacingCars: List<RacingCar>) {
 }
