+++ src/main/kotlin/me/parker/nextstep/kotlinracingcar/domain/RacingGameResult.kt
@@ -1,4 +1,4 @@
-package me.parker.nextstep.kotlinracingcar
+package me.parker.nextstep.kotlinracingcar.domain
 
 data class RacingGameResult(val rounds: List<RacingCarGameRound>, val winners: Winners) {
 }
