+++ src/main/kotlin/me/parker/nextstep/kotlinracingcar/domain/rule/RacingGameMoveRule.kt
@@ -1,4 +1,4 @@
-package me.parker.nextstep.kotlinracingcar.rule
+package me.parker.nextstep.kotlinracingcar.domain.rule
 
 interface RacingGameMoveRule {
     fun isForward(): Boolean
