+++ src/test/kotlin/me/parker/nextstep/kotlinracingcar/domain/RacingCarTest.kt
@@ -1,7 +1,7 @@
-package me.parker.nextstep.kotlinracingcar
+package me.parker.nextstep.kotlinracingcar.domain
 
-import me.parker.nextstep.kotlinracingcar.rule.RandomRacingGameMoveRule
-import me.parker.nextstep.kotlinracingcar.rule.TestRacingGameMoveRule
+import me.parker.nextstep.kotlinracingcar.domain.rule.RandomRacingGameMoveRule
+import me.parker.nextstep.kotlinracingcar.domain.rule.TestRacingGameMoveRule
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
