+++ src/test/kotlin/me/parker/nextstep/kotlinracingcar/domain/RacingGameResultTest.kt
@@ -1,6 +1,10 @@
-package me.parker.nextstep.kotlinracingcar
+package me.parker.nextstep.kotlinracingcar.domain
 
-import me.parker.nextstep.kotlinracingcar.rule.RandomRacingGameMoveRule
+import me.parker.nextstep.kotlinracingcar.domain.RacingCar
+import me.parker.nextstep.kotlinracingcar.domain.RacingCarGameRound
+import me.parker.nextstep.kotlinracingcar.domain.RacingGameResult
+import me.parker.nextstep.kotlinracingcar.domain.Winners
+import me.parker.nextstep.kotlinracingcar.domain.rule.RandomRacingGameMoveRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
