+++ src/test/kotlin/me/parker/nextstep/kotlinracingcar/domain/rule/TestRacingGameMoveRule.kt
@@ -1,4 +1,6 @@
-package me.parker.nextstep.kotlinracingcar.rule
+package me.parker.nextstep.kotlinracingcar.domain.rule
+
+import me.parker.nextstep.kotlinracingcar.domain.rule.RacingGameMoveRule
 
 class TestRacingGameMoveRule(private var isForward: Boolean) : RacingGameMoveRule {
